+++ packages/nimble-components/src/utilities/tests/matrix.ts
@@ -1,58 +1,29 @@
 import { html, repeat, ViewTemplate } from '@microsoft/fast-element';
-import { Theme } from '../../theme-provider/types';
 
-export const backgroundStates = [
-    {
-        name: `"${Theme.Light}" theme on white`,
-        value: '#F4F4F4',
-        theme: Theme.Light
+export const sharedMatrixParameters = () => ({
+    controls: {
+        hideNoControlsWarning: true
     },
-    {
-        name: `"${Theme.Color}" theme on dark green`,
-        value: '#044123',
-        theme: Theme.Color
+    a11y: { disabled: true },
+    docs: {
+        source: {
+            code: null
+        },
+        transformSource: (source: string): string => source
     },
-    {
-        name: `"${Theme.Dark}" theme on black`,
-        value: '#252526',
-        theme: Theme.Dark
+    backgrounds: {
+        disable: true,
+        grid: {
+            disable: true
+        }
+    },
+    viewMode: 'canvas',
+    previewTabs: {
+        'storybook/docs/panel': {
+            hidden: true
+        }
     }
-] as const;
-
-export type BackgroundState = typeof backgroundStates[number];
-
-export type DisabledState = [string, boolean];
-export const disabledStates: DisabledState[] = [
-    ['', false],
-    ['Disabled', true]
-];
-
-export type InvalidState = [string, string];
-export const invalidStates: InvalidState[] = [
-    ['', ''],
-    ['Invalid', 'invalid']
-];
-
-export type ReadOnlyState = [string, boolean];
-export const readOnlyStates: ReadOnlyState[] = [
-    ['', false],
-    ['Read-Only', true]
-];
-
-export type IconVisibleState = boolean;
-export const iconVisibleStates: IconVisibleState[] = [false, true];
-
-export type ExpandedState = [string, boolean];
-export const expandedStates: ExpandedState[] = [
-    ['Collapsed', false],
-    ['Expanded', true]
-];
-
-export type SelectedState = [string, boolean];
-export const selectedStates: SelectedState[] = [
-    ['Unselected', false],
-    ['Selected', true]
-];
+} as const);
 
 /**
  * Takes an array of state values that can be used with the template to match the permutations of the provided states.
@@ -184,39 +155,3 @@ export function createMatrix(
         `)}
     `;
 }
-
-/**
- * Wraps a given component template with a region for each of the available themes.
- */
-export const themeWrapper = (template: ViewTemplate): ViewTemplate => createMatrix(
-    ({ theme, value }: BackgroundState) => html`
-            <nimble-theme-provider theme="${theme}">
-                <div style="background-color: ${value}; padding:20px;">
-                    ${template}
-                </div>
-            </nimble-theme-provider>
-        `,
-    [backgroundStates]
-);
-
-// A customized theme wrapper (not themeWrapper like the other controls) so we can create different stories for each theme, rather
-// than having a single Theme Matrix story. This is useful for when the UI under test can't be tested multiple times on a
-// single page, but you still want to test the UI for each theme.
-export const singleThemeWrapper = (
-    template: ViewTemplate,
-    backgroundState: BackgroundState
-): ViewTemplate => html`
-    <nimble-theme-provider theme="${backgroundState.theme}">
-        <div
-            style="
-                background-color: ${backgroundState.value};
-                position: absolute;
-                width: 100%;
-                height: 100%;
-                left: 0px;
-                top: 0px;
-            "
-        ></div>
-        ${template}
-    </nimble-theme-provider>
-`;