+++ angular-workspace/projects/ni/nimble-angular/CHANGELOG.md
@@ -1,8 +1,56 @@
 # Change Log - @ni/nimble-angular
 
-This log was last generated on Fri, 22 Apr 2022 14:26:36 GMT and should not be manually modified.
+This log was last generated on Sat, 23 Apr 2022 05:05:02 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 5.0.10
+
+Sat, 23 Apr 2022 05:05:02 GMT
+
+### Patches
+
+- Bump @ni/nimble-components to v8.1.10
+
+## 5.0.9
+
+Sat, 23 Apr 2022 04:31:01 GMT
+
+### Patches
+
+- Bump @ni/nimble-components to v8.1.9
+
+## 5.0.8
+
+Sat, 23 Apr 2022 04:15:08 GMT
+
+### Patches
+
+- Bump @ni/nimble-components to v8.1.8
+
+## 5.0.7
+
+Fri, 22 Apr 2022 22:36:24 GMT
+
+### Patches
+
+- Bump @ni/nimble-components to v8.1.4
+
+## 5.0.6
+
+Fri, 22 Apr 2022 16:09:27 GMT
+
+### Patches
+
+- Bump @ni/nimble-components to v8.1.3
+
+## 5.0.5
+
+Fri, 22 Apr 2022 15:09:11 GMT
+
+### Patches
+
+- Bump @ni/nimble-components to v8.1.2
 
 ## 5.0.4