+++ packages/nimble-components/CHANGELOG.md
@@ -1,8 +1,56 @@
 # Change Log - @ni/nimble-components
 
-This log was last generated on Fri, 22 Apr 2022 14:26:36 GMT and should not be manually modified.
+This log was last generated on Sat, 23 Apr 2022 05:05:02 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 8.1.10
+
+Sat, 23 Apr 2022 05:05:02 GMT
+
+### Patches
+
+- Bump @ni/nimble-tokens to v3.0.13
+
+## 8.1.9
+
+Sat, 23 Apr 2022 04:31:01 GMT
+
+### Patches
+
+- Bump @ni/nimble-tokens to v3.0.12
+
+## 8.1.8
+
+Sat, 23 Apr 2022 04:15:08 GMT
+
+### Patches
+
+- Bump @ni/nimble-tokens to v3.0.11
+
+## 8.1.4
+
+Fri, 22 Apr 2022 22:36:24 GMT
+
+### Patches
+
+- Bump @ni/nimble-tokens to v3.0.7
+
+## 8.1.3
+
+Fri, 22 Apr 2022 16:09:27 GMT
+
+### Patches
+
+- Bump @ni/nimble-tokens to v3.0.6
+
+## 8.1.2
+
+Fri, 22 Apr 2022 15:09:11 GMT
+
+### Patches
+
+- Bump @ni/nimble-tokens to v3.0.4
 
 ## 8.1.1